% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve for Lab 6}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{a}{the ending bound}
}
\value{
returns a curve with a specific shaded area
}
\description{
myncurve for Lab 6
}
\examples{
myncurve(mu = 5, sigma = 3, a = 2)
}
