% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySample.R
\name{mysample}
\alias{mysample}
\title{Title
is a function that takes 3 arguments
it samples a vector of 1-10 values and displays a chart with time in between}
\usage{
mysample(n = 1, iter = 10, time = 0.5)
}
\arguments{
\item{n}{amount of choices}

\item{iter}{number of iterations}

\item{time}{amount of time between iterations}
}
\value{
returns mysample
}
\description{
Title
is a function that takes 3 arguments
it samples a vector of 1-10 values and displays a chart with time in between
}
\examples{
mysample(1, iter = 10, time = 0.5)
}
